Cant figure out the boat shake
I'm trying to reference all the children triggers with this.transform.Find so i can attach this script to the parent object because if i have the script attached to one trigger it doesnt update the others  when its ready to flip. I was able to get the boats to shake a variable amount of times depending on the boat but not able to flip because the script was attached to the trigger and not the parent. Also while I'm writing this out, all 3 of my animators just stopped working, i cant see or edit any of them.